requirements.txt
amqp==2.2.2
anyjson==0.3.3
backoff==1.4.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==2.1.2
boto3==1.5.12
celery==4.1.0
cryptography==2.1.4
django-autocomplete-light==2.3.3  # pyup: >=2.0.0,<3.0.0
django-ckeditor==5.4.0
django-constance==2.0.0
django-extensions==1.9.8
django-filter==1.1.0
django-grappelli==2.10.1
django-mailbox==4.6.0
django-picklefield==1.0.0
django-s3-storage==0.12.2
Django==1.11.7
djangorestframework-jwt==1.11.0
djangorestframework==3.7.7
docutils==0.14
drfdocs==0.0.11
ec2-metadata==1.6.0
elasticsearch==5.4.0
feedparser==5.2.1
geoip2==2.6.0
geopy==1.11.0
gunicorn==19.7.1
html5lib==1.0.1
httmock==1.2.6
httplib2==0.10.3
jira==1.0.10
kombu==4.1.0
langdetect==1.0.7
maxminddb==1.3.0
nltk==3.2.5
oauth2==1.9.0.post1
oauthlib==2.0.6
pika==0.11.2
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2017.3
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
requests==2.18.4
sauceclient==1.0.0
selenium==3.8.1
six==1.10.0
testfixtures==5.3.1
textblob==0.15.0
tlslite==0.4.9
tweepy==3.5.0
urllib3==1.22
